refactor: agent-server quality fixes and context directory consolidation
- Fix port collision: agent-server now defaults to 4802 (was 4801, same as
  chat-relay). Updated in index.ts, agent_server.rs (×2), and lib.rs.
- Replace qaAttempts=99 hack with qaPassed boolean in graph state. qaOutcome
  now routes on state.qaPassed instead of a magic number.
- Add test script, vitest ^4.1.0, vitest.config.ts, and private:true to
  agent-server package.json (all 9 existing tests now runnable via pnpm test).
- Consolidate context/ into contexts/: move ChatContext.tsx to match
  ExecutionContext.tsx location; update all 12 import paths.
