+++ apps/desktop/src/App.tsx
@@ -2,7 +2,7 @@ import { BrowserRouter, Routes, Route, Navigate } from "react-router-dom";
 import AppLayout from "./layouts/AppLayout";
 import PreferencesPage from "./pages/PreferencesPage";
 import { getDefaultSection } from "./lib/preferences";
-import { ChatProvider } from "./context/ChatContext";
+import { ChatProvider } from "./contexts/ChatContext";
 import { ObservatoryProvider } from "./hooks/useObservatoryData";
 import HarnessFilePage from "./pages/harness/HarnessFilePage";
 import PluginsPage from "./pages/harness/PluginsPage";
