+++ apps/desktop/src/components/chat/ChatInput.tsx
@@ -1,5 +1,5 @@
 import { useState, useRef, useCallback, useEffect } from "react";
-import { useChat } from "../../context/ChatContext";
+import { useChat } from "../../contexts/ChatContext";
 
 export default function ChatInput() {
   const { state, sendChat, sendTyping } = useChat();
