+++ packages/agent-server/package.json
@@ -1,12 +1,14 @@
 {
   "name": "@harness-kit/agent-server",
   "version": "0.1.0",
+  "private": true,
   "type": "module",
   "main": "dist/index.js",
   "scripts": {
     "build": "tsc",
     "dev": "tsx watch src/index.ts",
-    "start": "node dist/index.js"
+    "start": "node dist/index.js",
+    "test": "vitest run"
   },
   "dependencies": {
     "@langchain/core": "^1.1.0",
@@ -26,6 +28,6 @@
     "@types/node": "^22.19.17",
     "tsx": "^4.7.0",
     "typescript": "^5.4.0",
-    "vitest": "^1.6.0"
+    "vitest": "^4.1.0"
   }
 }
