+++ packages/agent-server/src/index.ts
@@ -1,7 +1,7 @@
 import { createServer } from './http.js';
 import { createWsServer } from './ws.js';
 
-const PORT = Number(process.env.AGENT_SERVER_PORT ?? 4801);
+const PORT = Number(process.env.AGENT_SERVER_PORT ?? 4802);
 const app = createServer();
 const server = app.listen(PORT, () => {
   console.log(`[agent-server] listening on :${PORT}`);
